Fixed some minor routing error.
I wanna note here, that adding a department to this "Projekt" is an
adventure. And not the fun kind of adventure, not even the adventure
where you might learn something out of it, but the kind of adventure
that makes you sad, and makes you curse about those students who call
themselfs "Programmers".